% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-DHARMa.R
\name{get_DHARMAres}
\alias{get_DHARMAres}
\title{Plot DHARMa Residuals for GLM2 and Survreg}
\usage{
get_DHARMAres(fit, compare_fit = NULL)
}
\arguments{
\item{fit}{A fitted model object (of class glm, glm2, or survreg).}

\item{compare_fit}{A fitted model object if model needs to be compared with a previous model}
}
\value{
A ggplot object.
}
\description{
This function creates a standardized diagnostic plot for DHARMa residuals
using ggplot2.
}
