% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-influ.R
\name{get_influ_glm}
\alias{get_influ_glm}
\title{Get the influence metric for GLM/survreg}
\usage{
get_influ_glm(fit, predictor = NULL)
}
\arguments{
\item{fit}{a model fit}
}
\value{
a \code{data.frame}
}
\description{
Get the influence metric for GLM/survreg
}
