% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{get_preds}
\alias{get_preds}
\title{Extract model predictions by term}
\usage{
get_preds(fit, raw_data = NULL)
}
\arguments{
\item{fit}{fitted model object}
}
\value{
A list containing preds dataframe and other stuff
}
\description{
Extract model predictions by term
}
