% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-index.R
\name{get_unstandardised}
\alias{get_unstandardised}
\title{Get the unstandardised indices}
\usage{
get_unstandardised(fit, year = NULL, rescale = 1, predictor = NULL)
}
\arguments{
\item{fit}{An object of class \code{brmsfit}.}

\item{year}{The year or time label (e.g. year, Year, fishing_year, etc).}

\item{rescale}{How to re-scale the series. Choose from "raw" to retain the raw unstandardised series, or a number to re-scale by.}

\item{predictor}{=1 or =2 if a hurdle model is used. If NULL, combined index will be returned.}
}
\value{
a \code{data.frame} or a \code{ggplot} object.
}
\description{
Get the unstandardised indices
}
