% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{mean_or_mode}
\alias{mean_or_mode}
\title{Impute values for new data}
\usage{
mean_or_mode(z)
}
\arguments{
\item{z}{a vector.}
}
\value{
Mean for numeric variables and mode for categorical
}
\description{
Impute values for new data
}
