% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-cdi.R
\name{plot_cdi}
\alias{plot_cdi}
\title{Plot Coefficient Distribution and Influence (CDI)}
\usage{
plot_cdi(preds_list, compare_preds_list = NULL)
}
\arguments{
\item{fit}{A fitted model object (e.g., from `glm` or `gam`).}

\item{year}{Character string. The name of the temporal column (e.g., "fyear")
to be used as the focus variable on the y-axis of the distribution plot.}
}
\description{
Generates a multi-pane layout visualising the coefficients, data distribution,
and relative influence of a model predictor.
}
