% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-cdi.R
\name{cdi_plot_with_indicators}
\alias{cdi_plot_with_indicators}
\title{Plot Coefficient Distribution and Influence (CDI) and calculate comparison traffic light metrics}
\usage{
cdi_plot_with_indicators(preds_list, compare_preds_list = NULL)
}
\arguments{
\item{fit}{A fitted model object (e.g., from `glm` or `gam`).}

\item{year}{Character string. The name of the temporal column (e.g., "fyear") 
to be used as the focus variable on the y-axis of the distribution plot.}

\item{predictor}{}
}
\description{
Generates a multi-pane layout visualising the coefficients, data distribution, 
and relative influence of a model predictor.
}
