% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-DHARMa.R
\name{plot_DHARMares}
\alias{plot_DHARMares}
\title{Plot DHARMa Residuals for GLM2 and Survreg}
\usage{
plot_DHARMares(fit)
}
\arguments{
\item{fit}{A fitted model object (of class glm, glm2, or survreg).}
}
\value{
A ggplot object.
}
\description{
This function creates a standardized diagnostic plot for DHARMa residuals
using ggplot2.
}
