% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-index.R
\name{plot_index}
\alias{plot_index}
\title{Plot the standardised and unstandardised indices}
\usage{
plot_index(
  index,
  component = "Combined",
  fill = "black",
  probs = c(0.25, 0.75),
  rescale = 1,
  predictor = NULL,
  show_unstandardised = TRUE
)
}
\arguments{
\item{index}{A dataframe with unstan and stan idex for the model.}

\item{fill}{the fill colour for the percentiles.}

\item{probs}{The percentiles to be computed by the \code{quantile} function.}

\item{rescale}{the index of the series to rescale to. If set to NULL then no rescaling is done.}

\item{show_unstandardised}{show the unstandardised series or not.}

\item{year}{the year or time label.}
}
\value{
a \code{ggplot} object.
}
\description{
In this plot the unstandardised indices is the geometric mean of the data.
}
