% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-resids.R
\name{plot_RIC}
\alias{plot_RIC}
\title{Plot Residual Implied Index}
\usage{
plot_RIC(fit, grouping_var = "stat_area")
}
\arguments{
\item{fit}{A fitted model object}

\item{grouping_var}{A character string specifying the column name to group by. 
Defaults to 'stat_area'.}
}
\value{
A ggplot2 object showing the implied vs. scaled indices.
}
\description{
Plot Residual Implied Index
}
