% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-index.R
\name{plot_sos}
\alias{plot_sos}
\title{Plot Status of the stock}
\usage{
plot_sos(
  cidx,
  CPUE_set,
  ref_period = NULL,
  period_type = "target",
  ref_series,
  landings_data = NULL,
  plot_exploitation = TRUE,
  cpue_smooth = FALSE,
  bmsy_proxy = 40,
  ref_mean_type = "geometric"
)
}
\arguments{
\item{cidx}{A list of CPUE index data frames generated with get_index().}

\item{CPUE_set}{Character vector of series names to include.}

\item{ref_period}{Numeric vector of years for the reference period.}

\item{period_type}{Reference type one of c('target', 'soft_limit', 'hard_limit').}

\item{ref_series}{Index of the reference series in `CPUE_set`.}

\item{landings_data}{Data frame of landings/removals.}

\item{plot_exploitation}{Logical; if TRUE, includes the exploitation rate panel.}

\item{cpue_smooth}{Logical; applies loess smoothing to the CPUE.}

\item{bmsy_proxy}{Numeric proxy value (default 40).}

\item{component}{Character vector of index components, one of c('BInomial', 'Positive', 'Combined').}

\item{ref_index}{The name of the reference index component, one of c('BInomial', 'Positive', 'Combined').}
}
\value{
A patchwork ggplot object.
}
\description{
Generates a multi-panel plot showing CPUE indices relative to reference levels (if exist), 
removals, and relative exploitation rates.
}
