% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-step.R
\name{plot_step}
\alias{plot_step}
\title{Plot Step Results from brms Model Fits}
\usage{
plot_step(step_df, compare_step_df = NULL)
}
\arguments{
\item{step_df}{A data frame generated by the \code{get_step()} function.}

\item{compare_step_df}{Optional. A second data frame generated by \code{get_step()} 
to be overlaid or compared against the primary \code{step_df}. Defaults to NULL.}
}
\value{
A \code{ggplot} object visualising the impact of each explanatory variable on CPUE standardisation.
}
\description{
Plot Step Results from brms Model Fits
}
