% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-step.R
\name{plot_step2}
\alias{plot_step2}
\title{This requires that all steps be run using brms and then provided as a list of model fits.}
\usage{
plot_step2(step_df)
}
\arguments{
\item{step_df}{a list of model fits in the order that you want to compare them}
}
\description{
This requires that all steps be run using brms and then provided as a list of model fits.
}
